Implement toQuery in MatchPredicate
Relates to https://github.com/crate/crate/pull/11669
